LayoutTests/storage/indexeddb/readonly-properties-expected.txt

Test IndexedDB readonly properties

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = window.indexedDB || window.webkitIndexedDB || window.mozIndexedDB;
PASS indexedDB == null is false
IDBDatabaseException = window.IDBDatabaseException || window.webkitIDBDatabaseException;
PASS IDBDatabaseException == null is false
IDBCursor = window.IDBCursor || window.webkitIDBCursor;
PASS IDBCursor == null is false
IDBKeyRange = window.IDBKeyRange || window.webkitIDBKeyRange;
PASS IDBKeyRange == null is false
indexedDB.open(name, description)
db = event.target.result
request = db.setVersion('1')
transaction = event.target.result;
Deleted all object stores.
objectStore = db.createObjectStore('foo');
trying to set readonly property objectStore.transaction
objectStore.transaction = this
PASS objectStore.transaction is still [object IDBTransaction]
PASS successfullyParsed is true

TEST COMPLETE
